added Exceptions

+++ src/main/java/com/springproject/commerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.springproject.commerce.services;
 import com.springproject.commerce.dto.ProductDto;
 import com.springproject.commerce.entities.Product;
 import com.springproject.commerce.repositories.ProductRepository;
+import com.springproject.commerce.services.exceptions.DataBaseException;
+import com.springproject.commerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,8 +25,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDto findById(Long id) {
-        Optional<Product> result = productRepository.findById(id);
-        Product product = result.get();
+        Product product = productRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDto(product);
     }
 
@@ -41,14 +46,26 @@ public class ProductService {
 
     @Transactional
     public ProductDto update(Long id, ProductDto productDto) {
-        Product entity = productRepository.getReferenceById(id);
-        copyDtoForEntity(productDto, entity);
-        entity = productRepository.save(entity);
-        return new ProductDto(entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoForEntity(productDto, entity);
+            entity = productRepository.save(entity);
+            return new ProductDto(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
-    @Transactional
+
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-    productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DataBaseException("Falha de Integridade Referencial");
+        }
     }
 
     private static void copyDtoForEntity(ProductDto productDto, Product entity) {
